\name{analyst.OPLS}
\alias{analyst.OPLS}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Analyst the peaks with PLS-DA or OPLS-DA
}
\description{
PLS, and OPLS classification
}
\usage{
analyst.OPLS(labels, data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{labels}{
A response vector.
}
  \item{data}{
The result of \code{getDataMatrix} or \code{fillPeaks} function
}
}
