\name{analyst.RF}
\alias{analyst.RF}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Analyst the peaks with random forest
}
\description{
random forest algorithm (based on Breiman and Cutler's original Fortran code) for classification.
}
\usage{
analyst.RF(labels, data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{labels}{
A response vector.
}
  \item{data}{
The result of \code{getDataMatrix} or \code{fillPeaks} function
}
}
