\name{getMS}
\alias{getMS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
get MS of a LC-MS data file.
}
\description{
get MS of a LC-MS data file.
}
\usage{
getMS(filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{filename}{
The path of a LC-MS data file.
}
}

\value{
a LIST, use
\item{path}{path of each LC-MS data file.}
\item{MS}{MS.}
}
