\name{getPseudospecturm}
\alias{getPseudospecturm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get a pseudospecturm of a detected substance
}
\description{
Get a pseudospecturm which of a detected substance
}
\usage{
getPseudospecturm(r.group, index, std = "maxo")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{r.group}{
the result of grouping step
}
  \item{index}{
the index of group.
}
  \item{std}{
only 'maxo' is supported by now
}
}