\name{KPIC-package}
\alias{KPIC-package}
\alias{KPIC}
\docType{package}
\title{
\packageTitle{KPIC}
}
\description{
\packageDescription{KPIC}
}
\details{
Package:	 KPIC

Type:	 Package

Version:	 2.0

Initial version:	 0.2.3

Initial date:	 2017.3

License:	 LGPL (>= 3)
}
\author{
HC Ji
}

\keyword{ package }