% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{peak.analyst.koplsda}
\alias{peak.analyst.koplsda}
\title{peak.analyst.koplsda}
\usage{
peak.analyst.koplsda(X,Y,A,oax,opt='GS',nrcvouter=20,nrcvinner=10,permutation=50,kernelParams=c(1,30,29))
}

\arguments{
\item{X}{- n x p matrix, where n is the number of samples and p is the number of variables.}

\item{Y}{- n x 1 matrix.}

\item{A}{- The number of Y-predictive components (integer)..}

\item{oax}{-The number of Y-orthogonal components (integer)..}

\item{opt}{- Optional setting to do kernel parameter optimisation, 'no' for no optimisation, uses kernelParams, or 'SA' for simulated annealing (default for gaussian kernel), 'GS' for gridsearch between two values as input in kernelParams (default for polynomial kernel).}

\item{nrcvouter}{- number of outer cross-validation rounds (integer, default = 20).}

\item{nrcvinner}{- number of inner cross-validation rounds (integer, default = 10).}

\item{permutation}{- number of permutation for the randomization test.}

\item{kernelParams}{- settings for the kernel parameter, leave empty for optimisation with SA, give three values c(start, end, nrsteps) for gridsearch, between which the gridsearch will be performed and the number of steps; note the kernel parameter can not be zero, and polynomial parameters can only be integer.}
}

\value{
\item{model}{Model calculated with all data and using single (i.e. not-nested) cross-validation optimisation of the kernel parameter.}

\item{cv}{Cross-validation results.}

\item{da}{Cross-validation results specifically for discriminant analysis (DA) cases (list).}

\item{permutation}{permutation testing results.}

}

\description{
Function for performing K-OPLS cross-validation for a set of Y-orthogonal components. The function returns a number of diagnostic parameters which can be used to determine the optimal number of model components. Optimisation of the kernel parameter is possible with grid search within defined limits and simulated annealing.
}

\references{
Bylesjo M, Rantalainen M, Nicholson JK, Holmes E and Trygg J. K-OPLS package: kernel-based orthogonal projections to latent structures for prediction and interpretation in feature space, BMC Bioinformatics 2008; 19;9:106.
}


