% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{peak.analyst.oplsda}
\alias{peak.analyst.oplsda}
\title{peak.analyst.oplsda}
\usage{
peak.analyst.oplsda(X, Y, ncomp, northo)
}

\arguments{
\item{X}{- n x p matrix, where n is the number of samples and p is the number of variables.}

\item{Y}{- n x 1 matrix.}

\item{ncomp}{- Integer: predictive componenents in case of OPLS to extract.}

\item{northo}{- Integer: number of orthogonal components for OPLS.}

}
\value{
An object of class 'opls' (see ropls package) and VIPs
}



