% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{peak.analyst.rf}
\alias{peak.analyst.rf}
\title{peak.analyst.rf}
\description{
random forest algorithm (based on Breiman and Cutler's original Fortran code) for classification.
}
\usage{
peak.analyst.rf(x, y, ntree=500, maxnodes=7)
}

\arguments{
\item{x}{- n x p matrix, where n is the number of samples and p is the number of variables.}

\item{y}{- n x 1 matrix.}

\item{ntree}{- Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times.}

\item{maxnodes}{- Maximum number of terminal nodes trees in the forest can have. If not given, trees are grown to the maximum possible (subject to limits by nodesize). If set larger than maximum possible, a warning is issued.}
}
\value{
An object of class randomForest (see randomForest package) and VIPs
}


