\name{PICeval}
\alias{PICeval}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
evaluate the peak shape of every PIC
}
\description{
evaluate the peak shape of every PIC with gauss fitness and sharpness
}
\usage{
PICeval(PICs)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{PICs}{
the result of getPIC function
}
}

\value{
the peakmat with another two column, which are gaussfitness and sharpness
}

\author{
HC Ji
}