\name{PICset.getPeaks}
\alias{PICset.getPeaks}

\title{
The \code{getPeaks} function for a set of samples.
}
\description{
This function is used to apply \code{PICfit} method to a PIC set.
}
\usage{
PICset.getPeaks(picset)
}

\arguments{
  \item{picset}{
The result object of \code{PICset}, \code{PICset.kmeans}, \code{PICset.split}, \code{PICset.resolve} or \code{PICset.fit} founction.
}
}

\value{
The processed \code{picset} object
}
