\name{pretreat}
\alias{pretreat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
pretreat the data matrix.
}
\description{
pretreat the data matrix.
}
\usage{
pretreat(r_fillpeaks,normalization='sum',scaling='auto scaling')
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{r_fillpeaks}{
the result of the fillpeaks function
}
  \item{normalization}{
the method of normalization. only 'sum' can be used for now.
}
  \item{scaling}{
the method of scaling. can be 'auto scaling' or 'center'
}
}
