\name{viewAlign}
\alias{viewAlign}

\title{
View the result of alignment.
}
\description{
View the result of alignment.
}
\usage{
viewAlign(groups_raw, groups_align)
}

\arguments{
  \item{groups_raw}{
The result of \code{PICgroup}
}
  \item{groups_align}{
The result of \code{PICalign}
}
}

\value{
a shiny app.
}
