\name{viewMS}
\alias{viewMS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
View MS.
}
\description{
View MS.
}
\usage{
viewMS(MS)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MS}{
The result object of \code{getMS} function.
}
}

\value{
A shiny app.
}
