\name{viewPICs}
\alias{viewPICs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
View the PICs.
}
\description{
View the PICs.
}
\usage{
viewPICs(pics)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pics}{
The result object of \code{getPIC}, \code{getPIC.kmeans}, \code{PICsplit}, \code{PICresolve} or \code{PICfit} founction.
}
}

\value{
A shiny app.
}
